fix: functions in correct archives
