+++ .env.example
@@ -28,5 +28,7 @@ EMAIL_PASSWORD=
 # JWT
 ACCESS_TOKEN_SECRET=
 REFRESH_TOKEN_SECRET=
+PASSWORD_TOKEN_SECRET=
 ACCESS_TOKEN_EXPIRE=
 REFRESH_TOKEN_EXPIRE=
+PASSWORD_TOKEN_EXPIRE=
