+++ src/routes/LegalPersonRoutes.js
@@ -1,5 +1,6 @@
 import { Router } from 'express';
 
+import * as BaseUserController from '../controllers/BaseUserController.js';
 import * as LegalPersonController from '../controllers/LegalPersonController.js';
 
 const LegalPersonRoutes = Router();
@@ -12,5 +13,7 @@ LegalPersonRoutes.route('/:_id')
   .get(LegalPersonController.getById)
   .put(LegalPersonController.update)
   .delete(LegalPersonController.destroy);
+
+LegalPersonRoutes.put('/:token', BaseUserController.verifyEmail);
 
 export default LegalPersonRoutes;
