+++ src/routes/PhysicalPersonRoutes.js
@@ -1,5 +1,6 @@
 import { Router } from 'express';
 
+import * as BaseUserController from '../controllers/BaseUserController.js';
 import * as PhysicalPersonController from '../controllers/PhysicalPersonController.js';
 
 const PhysicalPersonRoutes = Router();
@@ -12,5 +13,7 @@ PhysicalPersonRoutes.route('/:_id')
   .get(PhysicalPersonController.getById)
   .put(PhysicalPersonController.update)
   .delete(PhysicalPersonController.destroy);
+
+PhysicalPersonRoutes.put('/:token', BaseUserController.verifyEmail);
 
 export default PhysicalPersonRoutes;
