+++ src/validators/BaseUserValidator.js
@@ -1,0 +1,12 @@
+/* eslint-disable import/prefer-default-export */
+import { z } from 'zod';
+
+import validate from './validate.js';
+
+export const verifyEmail = validate(
+  z.object({
+    params: z.object({
+      token: z.string({ required_error: 'User token is required' }),
+    }),
+  })
+);
